# app/api/openapi.yaml
openapi: 3.0.3
info:
  title: SessionProxy Control Plane API
  description: >
    Owner-facing API for SessionProxy. This documents the contract
    implemented by hand in internal/transport/http (chi router); it is kept
    in sync manually rather than driving code generation, so treat this file
    as the source of truth for what the API promises, not as an artifact
    that regenerates the server.
  version: "1.0.0"
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /links:
    get:
      summary: List the caller's shared_links
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, terminated] }
      responses:
        "200": { description: OK }
    post:
      summary: Create a shared_link for an owned original_session (FR2)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "201": { description: Created }
        "409": { description: original_session is not active }
...
  /links/{linkID}/terminate:
    post:
      summary: Manually terminate a link (FR6/FR8)
      parameters: [{ $ref: "#/components/parameters/linkID" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string }
      responses:
        "204": { description: Terminated }
        "409": { description: Already terminated }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Either an owner JWT from /auth/login, or a long-lived API key
        (prefixed spk_) from POST /api-keys.
  parameters:
    linkID: { name: linkID, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    CreateLinkRequest:
      type: object
      required: [original_session_id]
      properties:
        original_session_id: { type: string, format: uuid }
        label: { type: string }
        ttl_seconds: { type: integer, format: int64 }
        policy_ids:
          type: array
          items: { type: string, format: uuid }
        endpoint_ids:
          type: array
          items: { type: string, format: uuid }